+++ lib/inline_forms/form_elements/time_helper.rb
@@ -10,9 +10,11 @@ module InlineForms
         link_to_inline_edit object, attribute, object.send(attribute).nil? ? "<i class='fi-plus'></i>".html_safe : object.send(attribute).to_datetime.strftime("%l:%M%P"), from_callee: __callee__
       end
     
+      # Native <input type="time"> (8.1.25; replaces the jQuery timepicker).
+      # Submits 24h "HH:MM", which Active Record casts directly to a time column.
       def time_select_edit(object, attribute)
         css_id = 'timepicker_' + object.class.to_s.underscore + '_' + object.id.to_s + '_' + attribute.to_s
-        out = text_field_tag attribute, ( object.send(attribute).nil? ? "" : object.send(attribute).to_datetime.strftime("%l:%M%P") ), :id => css_id, :class =>'timepicker'
+        time_field_tag attribute, ( object.send(attribute).nil? ? "" : object.send(attribute).to_datetime.strftime("%H:%M") ), :id => css_id, :class => 'time_select'
       end
     
       def time_select_update(object, attribute)